Simplify and harden the SimpleDAE reference FMU
- Restructure reference-FMUs/ to mirror the Reference-FMUs Dahlquist
  layout: one model.c, shared fmi3Functions.c and model_instance.c
  (renamed from cosimulation.c, doFixedStep removed), FMI 3.0 ME only
- Add CMakeLists.txt with C99-strict build, silent ZIP packaging, and
  install message; add multi-platform GitHub Actions build matrix
- Update fmi3LayeredStandardDaeManifest.xsd to accept
  <fmiLayeredStandardManifest> with <fmiModelDescription> containing
  <AlgebraicVariables> and <ModelStructure>/<Residual>/<Formulation>
- Fix fmi-ls-manifest.xml: correct fmi-ls-name, description, schema
  pointer, and convert Residual to use Formulation child elements
- Add XML dependency comments to modelDescription.xml and manifest
- Add FMPy ODE-mode simulation test and DAE-mode placeholder
- Add nightly.link download badge to README
- Fix docs: manifest table caption, attribute values, version
