+++ .github/workflows/validate-xml.yml
@@ -4,10 +4,12 @@ on:
   push:
     paths:
       - 'examples/**/*.xml'
+      - 'reference-FMUs/**/*.xml'
       - 'schema/**'
   pull_request:
     paths:
       - 'examples/**/*.xml'
+      - 'reference-FMUs/**/*.xml'
       - 'schema/**'
 
 jobs:
@@ -16,16 +18,26 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
-        uses: actions/checkout@v6
+        uses: actions/checkout@v7
 
-      - name: Validate XSD manifest
-        uses: phpcsstandards/xmllint-validate@v1
-        with:
-          pattern: schema/fmi3LayeredStandardDaeManifest.xsd
-          xsd-url: https://www.w3.org/2012/04/XMLSchema.xsd
+      - name: Install xmllint
+        run: sudo apt-get install -y --no-install-recommends libxml2-utils
 
-      - name: Validate XML examples
-        uses: phpcsstandards/xmllint-validate@v1
-        with:
-          pattern: 'examples/**/*.xml'
-          xsd-file: 'schema/fmi3LayeredStandardDaeManifest.xsd'
+      - name: Download FMI 3.0 schemas
+        run: |
+          mkdir fmi3-schema
+          for f in fmi3Annotation fmi3AttributeGroups fmi3BuildDescription fmi3InterfaceType \
+                   fmi3LayeredStandardManifest fmi3ModelDescription fmi3Terminal \
+                   fmi3TerminalsAndIcons fmi3Type fmi3Unit fmi3Variable fmi3VariableDependency; do
+            curl -sSfL "https://raw.githubusercontent.com/modelica/fmi-standard/v3.0.2/schema/${f}.xsd" \
+              -o "fmi3-schema/${f}.xsd"
+          done
+
+      - name: Validate DAE manifest XSD (well-formed)
+        run: xmllint --noout schema/fmi3LayeredStandardDaeManifest.xsd
+
+      - name: Validate fmi-ls-manifest.xml files
+        run: find . -name "fmi-ls-manifest.xml" | xargs xmllint --noout --schema schema/fmi3LayeredStandardDaeManifest.xsd
+
+      - name: Validate modelDescription.xml files
+        run: find reference-FMUs -name "modelDescription.xml" | xargs xmllint --noout --schema fmi3-schema/fmi3ModelDescription.xsd
