+++ .gitignore
@@ -7,6 +7,8 @@ docs/stem-*.svg
 # PyCharm projects
 .idea/
 .venv/
+.pytest_cache/
+__pycache__/
 
 # schema figure generator
 xsddiagram/
