+++ examples/xml/fmi-ls-manifest.xml
@@ -1,9 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<fmi-dae
+<fmiLayeredStandardManifest
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:noNamespaceSchemaLocation="../../schema/fmi3LayeredStandardDaeManifest.xsd"
+  xsi:schemaLocation="http://fmi-standard.org/fmi-ls-manifest ../../schema/fmi3LayeredStandardDaeManifest.xsd"
+  xmlns="http://fmi-standard.org/fmi-ls-manifest"
   xmlns:fmi-ls="http://fmi-standard.org/fmi-ls-manifest"
   fmi-ls:fmi-ls-name="org.fmi-standard.fmi-ls-dae"
   fmi-ls:fmi-ls-version="0.1"
   fmi-ls:fmi-ls-description="Layered standard for DAE support in FMI."
-  isDaeFmu="true"/>
+  isDaeFmu="true">
+</fmiLayeredStandardManifest>
