--- reference-FMUs/SimpleDAE/DAE/extra/org.fmi-standard.fmi-ls-dae/fmi-ls-manifest.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<fmiLayeredStandardManifest
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://fmi-standard.org/fmi-ls-manifest ../../schema/fmi3LayeredStandardManifest.xsd"
-  xmlns="http://fmi-standard.org/fmi-ls-manifest"
-  xmlns:fmi-ls="http://fmi-standard.org/fmi-ls-manifest"
-  fmi-ls:fmi-ls-name="fmi-ls-dae"
-  fmi-ls:fmi-ls-version="1.0"
-  fmi-ls:fmi-ls-description="Layered standard for DAE support in FMI">
-  <AlgebraicVariables>
-    <AlgebraicVariable
-      valueReference="5"/>
-    <AlgebraicVariable
-      valueReference="6"/>
-  </AlgebraicVariables>
-  <ModelStructure>
-    <ContinuousStateDerivative
-      valueReference="2"
-      dependencies="1 5 6 7"
-      dependenciesKind="dependent dependent dependent dependent"/>
-    <ContinuousStateDerivative
-      valueReference="4"
-      dependencies="1 3 5 6 7 8"
-      dependenciesKind="dependent dependent dependent dependent dependent dependent"/>
-    <Residual
-      valueReference="11"
-      dependencies="1 5 7 8"
-      dependenciesKind="dependent dependent dependent dependent"/>
-    <Residual
-      valueReference="12"
-      dependencies="3 5 6 7"
-      dependenciesKind="dependent dependent dependent dependent"/>
-    <Output
-      valueReference="9"
-      dependencies="1 5 7 8"
-      dependenciesKind="dependent dependent dependent dependent"/>
-    <Output
-      valueReference="10"
-      dependencies="3 6 7 8"
-      dependenciesKind="dependent dependent dependent dependent"/>
-  </ModelStructure>
-</fmiLayeredStandardManifest>
