--- reference-FMUs/SimpleDAE/DAE/sources/all.c
@@ -1,5 +1,0 @@
-#define FMI_VERSION 2
-
-#include "fmi2Functions.c"
-#include "model.c"
-#include "cosimulation.c"
