--- reference-FMUs/SimpleDAE/DAE/sources/buildDescription.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<fmiBuildDescription fmiVersion="3.0">
-
-  <BuildConfiguration modelIdentifier="DAE">
-    <SourceFileSet language="C99">
-      <SourceFile name="fmi3Functions.c"/>
-      <SourceFile name="model.c"/>
-      <SourceFile name="cosimulation.c"/>
-      <PreprocessorDefinition name="FMI_VERSION" value="3"/>
-    </SourceFileSet>
-  </BuildConfiguration>
-
-</fmiBuildDescription>
