--- reference-FMUs/SimpleDAE/DAE/sources/cosimulation.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "model.h"
-
-Status doFixedStep(ModelInstance *comp, bool* stateEvent, bool* timeEvent);
