+++ reference-FMUs/SimpleDAE/buildDescription.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<fmiBuildDescription fmiVersion="3.0">
+
+  <BuildConfiguration modelIdentifier="SimpleDAE">
+    <SourceFileSet language="C99">
+      <SourceFile name="fmi3Functions.c"/>
+      <SourceFile name="model_instance.c"/>
+      <SourceFile name="model.c"/>
+      <PreprocessorDefinition name="FMI_VERSION" value="3"/>
+      <PreprocessorDefinition name="DISABLE_PREFIX"/>
+    </SourceFileSet>
+  </BuildConfiguration>
+
+</fmiBuildDescription>
