+++ reference-FMUs/SimpleDAE/fmi-ls-manifest.xml
@@ -1,0 +1,59 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<fmiLayeredStandardManifest
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://fmi-standard.org/fmi-ls-manifest ../../schema/fmi3LayeredStandardDaeManifest.xsd"
+  xmlns="http://fmi-standard.org/fmi-ls-manifest"
+  xmlns:fmi-ls="http://fmi-standard.org/fmi-ls-manifest"
+  fmi-ls:fmi-ls-name="org.fmi-standard.fmi-ls-dae"
+  fmi-ls:fmi-ls-version="0.1"
+  fmi-ls:fmi-ls-description="Layered standard for DAE support in FMI."
+  isDaeFmu="true">
+
+  <AlgebraicVariables>
+    <AlgebraicVariable valueReference="5"/> <!-- z1 -->
+    <AlgebraicVariable valueReference="6"/> <!-- z2 -->
+  </AlgebraicVariables>
+  <ModelStructure>
+      <!-- der(x1) = sin(x1) + sin(z1*z2*u1)
+           DAE-mode deps: x1(vr=1), z1(vr=5), z2(vr=6), u1(vr=7) -->
+      <ContinuousStateDerivative
+        valueReference="2"
+        dependencies="1 5 6 7"
+        dependenciesKind="dependent dependent dependent dependent"/>
+      <!-- der(x2) = sin(x1*x2) + sin(z1*z2*u2) + u1^2
+           DAE-mode deps: x1(vr=1), x2(vr=3), z1(vr=5), z2(vr=6), u1(vr=7), u2(vr=8) -->
+      <ContinuousStateDerivative
+        valueReference="4"
+        dependencies="1 3 5 6 7 8"
+        dependenciesKind="dependent dependent dependent dependent dependent dependent"/>
+      <!-- __residual0 = z1*u1^2 + tanh(3*z1) + u2*x1^3
+           DAE-mode deps: x1(vr=1), z1(vr=5), u1(vr=7), u2(vr=8) -->
+      <Residual>
+        <Formulation
+          valueReference="11"
+          dependencies="1 5 7 8"
+          dependenciesKind="dependent dependent dependent dependent"/>
+      </Residual>
+      <!-- __residual1 = exp(z1*z2*u1)/3 - sin(z2*x2)
+           DAE-mode deps: x2(vr=3), z1(vr=5), z2(vr=6), u1(vr=7) -->
+      <Residual>
+        <Formulation
+          valueReference="12"
+          dependencies="3 5 6 7"
+          dependenciesKind="dependent dependent dependent dependent"/>
+      </Residual>
+      <!-- y1 = u1*x1*z1 + sin(u2)
+           DAE-mode deps: x1(vr=1), z1(vr=5), u1(vr=7), u2(vr=8) -->
+      <Output
+        valueReference="9"
+        dependencies="1 5 7 8"
+        dependenciesKind="dependent dependent dependent dependent"/>
+      <!-- y2 = u2*x2*z2 + sin(u1*u2)
+           DAE-mode deps: x2(vr=3), z2(vr=6), u1(vr=7), u2(vr=8) -->
+      <Output
+        valueReference="10"
+        dependencies="3 6 7 8"
+        dependenciesKind="dependent dependent dependent dependent"/>
+  </ModelStructure>
+
+</fmiLayeredStandardManifest>
